requirements.txt
azure-functions
airtable-python-wrapper==0.15.3
bcrypt==3.2.2
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.12
idna==3.3
pycparser==2.21
requests==2.27.1
urllib3==1.26.9
sendgrid==6.9.7
starkbank-ecdsa==2.0.3
